requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
distro==1.9.0
Flask==3.0.3
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
MarkupSafe==3.0.3
multidict==6.7.1
openai==2.26.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-dotenv==1.2.2
requests==2.32.5
sniffio==1.3.1
tqdm==4.67.3
twilio==9.3.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
watchdog==6.0.0
Werkzeug==3.1.6
yarl==1.23.0
